1645 - update label and description, refactor
