+++ server/libs/modules/components/baserow/src/test/java/com/bytechef/component/baserow/action/BaserowCreateRowActionTest.java
@@ -54,7 +54,6 @@ class BaserowCreateRowActionTest {
             .thenReturn(mockedExecutor);
         when(mockedExecutor.queryParameter(queryNameArgumentCaptor.capture(), queryValueArgumentCaptor.capture()))
             .thenReturn(mockedExecutor);
-
         when(mockedExecutor.body(bodyArgumentCaptor.capture()))
             .thenReturn(mockedExecutor);
         when(mockedExecutor.configuration(any()))
